refactor: rename the old irrelevant `Card` to `DEPRECATED_Card`.
