--- shared/ui/Card/index.tsx
@@ -1,1 +1,0 @@
-export { Card } from './Card';
