+++ shared/ui/DEPRECATED_Card/DEPRECATED_Card.module.scss
