+++ shared/ui/DEPRECATED_Card/DEPRECATED_Card.tsx
@@ -2,9 +2,9 @@ import { ReactNode } from 'react';
 
 import { clsx } from 'clsx';
 
-import s from './Card.module.scss';
+import s from './DEPRECATED_Card.module.scss';
 
-export const Card = (props: PropsType) => {
+export const DEPRECATED_Card = (props: PropsType) => {
   return <div className={clsx(s.component, props.className)}>{props.children}</div>;
 };
