+++ shared/ui/DEPRECATED_Card/index.tsx
@@ -1,0 +1,1 @@
+export * from './DEPRECATED_Card';
